% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api-archive.R
\name{archive_export}
\alias{archive_export}
\title{Export remake contents}
\usage{
archive_export(target_names = NULL, dependencies = TRUE, verbose = TRUE,
  require_current = TRUE, archive_file = "remake.zip",
  remake_file = "remake.yml")
}
\arguments{
\item{target_names}{Names of targets to export.}

\item{dependencies}{Export the \emph{dependencies} of
\code{target_names}?  The default is \code{TRUE}, which allows
targets such as \code{all} to be specified in order to export
everything that is a dependency of \code{all}.  If
\code{dependencies} is \code{FALSE}, all elements of
\code{target_names} must represent files or objects.}

\item{verbose}{Be verbose when reading the remake file?}

\item{require_current}{Logical indicating if the targets must be
up-to-date to be fetched.  If this is \code{TRUE} and the targets
are not up-to-date, then an error will be thrown.}

\item{archive_file}{Name of the archive file to generate, by
default \code{remake.zip}.}

\item{remake_file}{Remake file to read, by default
\code{remake.yml}.}
}
\value{
Invisibly, the name of the archive file generated.
However, this function is primarily useful for its side effect,
which is generating the archive.
}
\description{
Export an archive of remake contents.  Implicit files (those that
remake does not generate) are not exported.
}
\section{Warning}{

The archive format is subject to change and is
not suitable for long-term archiving.  Moreover, it depends on R's
internal rds format.  This format is itself not guaranteed to stay
constant, though it has for a long time now (see
\code{\link{serialize}}).  However, this is likely to be
reasonable for data interchange between computers or for
short/medium term export of results.  Until a lossless
representation of all R objects exists, the rds problem is not
likely to go away.
}

