% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{auto_gitignore}
\alias{auto_gitignore}
\title{Automatically generate .gitignore}
\usage{
auto_gitignore(remake_file = "remake.yml", check_git = TRUE,
  dry_run = FALSE)
}
\arguments{
\item{remake_file}{Name of the remake file to use, by default
\code{remake.yml}.}

\item{check_git}{Use the output of \code{git check-ignore} to
determine which files are already ignored?  This is only done if
git is detected to be installed and if the working directory
appears to be in a git repository.}

\item{dry_run}{Don't modify the .gitignore, but instead return a
character vector of what \emph{would} be added.}
}
\description{
Attempts to add targets that remake will generate to your
\code{.gitignore}.  If the file already exists, then the files
will be added (if not already present), otherwise a file will be
created.  If \code{check_git} is \code{TRUE} (the default) then we
attempt to check with \code{git} to see if the files are already
ignored by \emph{any} gitignore (including a system-specific
global gitignore) and only files that are not already excluded
will be added.  Probably after running this function you will want
to do some editing.
}

