% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_bindings}
\alias{create_bindings}
\alias{delete_bindings}
\title{Load remake bindings into the global environment}
\usage{
create_bindings(remake_file = "remake.yml")

delete_bindings(remake_file = "remake.yml")
}
\arguments{
\item{remake_file}{Name of the remakefile to read.  By default
\code{"remake.yml"}.}
}
\description{
Load bindings from \code{remake} into the global environment
(\code{\link{.GlobalEnv}}.  The resulting objects are "active
bindings" that when accessed will trigger a build of an object.
Conversely, \code{delete_bindings} undoes this and deletes the
bindings that \code{remake} made in the first place.
}

