% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diagram}
\alias{diagram}
\title{Make a figure with the dependency graph}
\usage{
diagram(m, targets = NULL)
}
\arguments{
\item{m}{A maker object}

\item{targets}{Optional character vector of targets.  If
specified, then only targets that these target depends on are
included.  Useful for filtering the graph.}
}
\description{
Plot the graph that maker generates.
}
\details{
This is really just a placeholder, but I want this here early as
an indication of where the package is headed.  Plus this is
something I have always wanted in make.  Current version is not
tunable on purpose.

This function \emph{will} change.
}

