% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{dump_environment}
\alias{dump_environment}
\title{Dump remake contents to environment}
\usage{
dump_environment(envir = .GlobalEnv, verbose = TRUE,
  allow_missing_packages = FALSE, remake_file = "remake.yml")
}
\arguments{
\item{envir}{Environment to copy into; by default the global environment.}

\item{verbose}{Be verbose when loading the remakefile}

\item{allow_missing_packages}{Allow missing packages when loading
remake file?}

\item{remake_file}{Name of the remakefile (by default
\code{remake.yml})}
}
\description{
Dump the contents of remake into an environment; by default the
global environment.  This is similar in effect to
\code{\link{create_bindings}} but does not create links; instead a
copy of everything that remake has built, plus all functions
sources into remake, are \emph{copied} into the environment.
}

