% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extensions}
\alias{extensions}
\title{Vector of file extensions}
\usage{
extensions()
}
\description{
Returns the vector of known file extensions.  If a target ends in
one of these, then it will be considered a file, rather than an
object.
}

