% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api.R
\name{fetch}
\alias{fetch}
\title{Fetch last computed}
\usage{
fetch(target_name, require_current = FALSE, remake_file = "remake.yml")
}
\arguments{
\item{target_name}{The name of a single target to fetch the value
of}

\item{remake_file}{Name of the remakefile (by default
\code{remake.yml})}
}
\value{
An R object.
}
\description{
Fetch the last computed value from the remake database.
}
\details{
The last computed value would be returned invisibly by
\code{make}, but this function provides a way of accessing values
without ever triggering a rebuild.  As such, it's possible that the
target is not made, or is not current, so there are options for
controlling what to do in this case.

It is an error to use this function with file targets (but see
\code{\link{is_current}} for checking currentness) and
\code{fetch_archive} for extracting files from archives.
}

