% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{file_extensions}
\alias{file_extensions}
\title{Vector of file extensions}
\usage{
file_extensions()
}
\description{
Returns the vector of \emph{default} file extensions.  If a target
ends in one of these, then it will be considered a file, rather
than an object.
}
\details{
To include \emph{additional} file extensions, include them in the
yaml like (at the top level):

\preformatted{
file_extensions: ["phy", "tre"]
}

Any number of extensions can be listed.  Don't use a leading
period (it will be dropped).  The union of file extensions listed
here and in \code{file_extensions()} will be used, so if you
accidently include a default extension (or if one is included in a
future remake version) it is no problem.
}

