% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_missing_packages}
\alias{install_missing_packages}
\title{Install missing packages}
\usage{
install_missing_packages(maker_file = "maker.yml", instructions = FALSE,
  missing_only = FALSE, skip_target_packages = FALSE)
}
\arguments{
\item{maker_file}{Name of the makerfile to look for the list of
required packages}

\item{instructions}{Rather than install anything, prints
instructions on how to install missing things}

\item{missing_only}{Install \emph{everything}, rather than just
missing packages.}

\item{skip_target_packages}{Skip packages that are mentioned only
in targets?}
}
\description{
Install missing packages
}
\author{
Rich FitzJohn
}

