% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-archive.R
\name{is_archive}
\alias{is_archive}
\title{Test if a zip file is likely to be a remake archive, as
created by \code{archive_create}}
\usage{
is_archive(archive_file = "remake.zip")
}
\arguments{
\item{archive_file}{Name of a file, by default \code{remake.zip}
(the default for \code{\link{archive_export}}.}
}
\description{
Test if a file is a remake archive.
}
