% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api-archive.R
\name{list_archive}
\alias{list_archive}
\title{List contents of a remake archive}
\usage{
list_archive(archive_file = "remake.zip", detail = FALSE)
}
\arguments{
\item{archive_file}{Name of the zip file to read from, by default
"remake.zip".}

\item{detail}{Return a data frame with more detail?}
}
\value{
A character vector with the contents of the archive.
}
\description{
List contents of a remake archive
}
\details{
NOTE: At present this unzips the entire archive.
}

