% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{list_targets}
\alias{list_dependencies}
\alias{list_targets}
\title{List targets}
\usage{
list_targets(remake_file = "remake.yml", type = NULL,
  include_implicit_files = FALSE, include_cleanup_targets = FALSE)

list_dependencies(target_names, type = NULL, include_implicit_files = FALSE,
  include_cleanup_targets = FALSE, remake_file = "remake.yml")
}
\arguments{
\item{remake_file}{Name of the remakefile to read (by default
\code{remake.yml})}

\item{type}{Type of target to return.  May be \code{fake},
\code{object}, \code{file} or \code{cleanup}.  Eventually subtypes
will be supported (knitr and plot targets) but this is not
possible yet.}

\item{include_implicit_files}{Logical scalar indicating if
implicit targets should be included.}

\item{include_cleanup_targets}{Logical scalar indicating if cleanup
targets (which are automatically generated) should be included.}

\item{target_names}{Names of targets to list dependencies of (for
\code{list_dependencies}).  These dependencies will be filtered as
for \code{list_targets}.  Dependencies are listed in topological
order: targets have no dependencies that occur later than them in
the vector.}
}
\value{
A character vector containing names of targets.
}
\description{
List targets contained within a remakefile
}
\details{
Do not rely on argument ordering here: please use argument names.
}

