% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{make}
\alias{make}
\title{Make one or more targets}
\usage{
make(target_names = NULL, ..., verbose = TRUE,
  allow_missing_packages = FALSE, remake_file = "remake.yml")
}
\arguments{
\item{target_names}{Character vector of names of targets to build,
or \code{NULL} to build the default target (if specified in the
remakefile).}

\item{...}{Additional future arguments, ignored for now.
Practically this means that all other arguments must be specified
by full name.}

\item{verbose}{Controls whether remake is verbose or not.  The
default (\code{TRUE}) prints out the name of each
target as it is built/checked.  This argument is passed to
\code{\link[=remake_verbose]{remake_verbose()}}; valid options are \code{TRUE},
\code{FALSE} and also the result of calling
\code{\link[=remake_verbose]{remake_verbose()}}.}

\item{allow_missing_packages}{Allow missing packages when loading
remake file?}

\item{remake_file}{Name of the remakefile (by default
\code{remake.yml}).}
}
\description{
Run remake to build one or more targets.
}
