% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{make_environment}
\alias{make_environment}
\title{Construct environment}
\usage{
make_environment(target_names = character(0), dependencies = FALSE,
  copy_functions = TRUE, verbose = TRUE, allow_missing_packages = FALSE,
  remake_file = "remake.yml")
}
\arguments{
\item{target_names}{Vector of target names to export.  If omitted,
then no targets are copied, though functions are still copied.}

\item{dependencies}{Should the dependenciesof \code{target_names}
also be copied over?  Setting this to \code{TRUE} is equivalent to
\code{make_environment(list_dependencies(target_names, type="file"))}
but shorter to type.}

\item{copy_functions}{Should functions be directly copied into
the retuned environment?  If \code{FALSE}, then the returned
environment has an environment with functions as its
\emph{parent}.  This is the same as the environment used by
\code{remake} so don't assign anything in here!  (This may change
if it ends up being a point of fragility.)}

\item{verbose}{Be verbose?}

\item{allow_missing_packages}{Allow missing packages when loading
remake file?}

\item{remake_file}{Remake file to use, by default
\code{remake.yml}.}
}
\description{
Construct an environment with remake target products, useful for
debugging.  Once you have an environment, you can
\code{\link{attach}} it (yes, it \emph{is} useful for something,
but be careful to \code{detach} later), extract elements or browse
it.
}

