% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/script.R
\name{make_script}
\alias{make_script}
\title{Make a single target}
\usage{
make_script(target_names = NULL, remake_file = "remake.yml")
}
\arguments{
\item{target_names}{Vector of names of targets to build, or
\code{NULL} to build the default target (if specified in the
remakefile).}

\item{remake_file}{Name of the remakefile (by default
\code{remake.yml}).  This is passed to \code{remake()}.}
}
\description{
Make a single target
}

