% Generated by roxygen2 (4.0.2): do not edit by hand
\name{maker_environment_browse}
\alias{maker_attach}
\alias{maker_detach}
\alias{maker_environment_browse}
\title{Interact with "maker environment" objects}
\usage{
maker_environment_browse(e, verbose = TRUE)

maker_attach(e, verbose = TRUE)

maker_detach(warn = TRUE, verbose = TRUE)
}
\arguments{
\item{e}{A \code{maker_environment} object.}

\item{verbose}{Logical indicating if information will be printed.
Because these functions to odd things to your search path, this is
\code{TRUE} by default.}

\item{warn}{Logical, indicating if a warning should be given when
there are no maker environments to detach.}
}
\description{
These functions may be useful in debugging workflows.  A
\code{maker_environment} object is created by running
\code{m$make_dependencies(target_name)}, where \code{target_name}
is the name of some target.  This returns an environment with all
the dependencies of that target so you can troubleshoot/develop
the rule that will process it.  The \emph{parent} environment of
this environment contains all the rules that maker knows about
(loaded from the files in the \code{sources} section of the
makerfile).  These functions ease interactions with these objects.
}

