% Generated by roxygen2 (4.0.2): do not edit by hand
\name{maker_install_packages}
\alias{maker_install_packages}
\title{Install packages required by makerfile}
\usage{
maker_install_packages(maker_file = "maker.yml")
}
\arguments{
\item{maker_file}{Name of the makerfile (by default
\code{maker.yml}).  This is passed to \code{maker()}.}
}
\description{
Install all missing packages that are required to use a
makerfile.  This is exactly equivalent to running
\code{maker::maker(maker_file)$install_packages()}, but is less
typing if you are not going to reuse the maker object.
}
\details{
No version comparison is done - see packrat for a more complete
package management solution.
}

