% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remake.R
\name{remake}
\alias{remake}
\title{Create a remake object}
\usage{
remake(remake_file = "remake.yml", verbose = TRUE, allow_cache = TRUE)
}
\arguments{
\item{remake_file}{Name of the remakefile (by default
\code{remake.yml})}

\item{verbose}{Controls whether remake is verbose or not.  By
default it is (\code{TRUE}), which prints out the name of each
target as it is built/checked.  This argument is passed to
\code{\link{remake_verbose}}; valid options are \code{TRUE},
\code{FALSE} and also the result of calling \code{remake_verbose}.}

\item{allow_cache}{Allow cached remake instances to be loaded?}
}
\description{
Creates a remake instance to interact with.
}
\details{
You probably should no longer interact with this: it will
disappear in a version or two.
}

