% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/active_bindings.R
\name{remake_create_bindings}
\alias{remake_create_bindings}
\alias{remake_delete_bindings}
\title{Load remake bindings into the global environment}
\usage{
remake_create_bindings(remake_file = "remake.yml")

remake_delete_bindings(remake_file = "remake.yml")
}
\arguments{
\item{remake_file}{Name of the remakefile to read.  By default
\code{"remake.yml"}.}
}
\description{
Load bindings from \code{remake} into the global environment
(\code{\link{.GlobalEnv}}.  The resulting objects are "active
bindings" that when accessed will trigger a build of an object.
}

