% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remake_environment.R
\name{remake_environment_browse}
\alias{remake_attach}
\alias{remake_detach}
\alias{remake_environment_browse}
\title{Interact with "remake environment" objects}
\usage{
remake_environment_browse(e, verbose = TRUE)

remake_attach(e, verbose = TRUE)

remake_detach(warn = TRUE, verbose = TRUE)
}
\arguments{
\item{e}{A \code{remake_environment} object.}

\item{verbose}{Logical indicating if information will be printed.
Because these functions to odd things to your search path, this is
\code{TRUE} by default.}

\item{warn}{Logical, indicating if a warning should be given when
there are no remake environments to detach.}
}
\description{
These functions may be useful in debugging workflows.  A
\code{remake_environment} object is created by running
\code{make_dependencies(m, target_name)}, where \code{target_name}
is the name of some target.  This returns an environment with all
the dependencies of that target so you can troubleshoot/develop
the rule that will process it.  The \emph{parent} environment of
this environment contains all the rules that remake knows about
(loaded from the files in the \code{sources} section of the
remakefile).  These functions ease interactions with these objects.
}

