% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/script.R
\name{source_maker_script}
\alias{source_maker_script}
\title{Source a maker script}
\usage{
source_maker_script(src, ..., envir = .GlobalEnv)
}
\arguments{
\item{src}{Contents}

\item{...}{Additional arguments passed to \code{\link{source}},
\emph{except} for the \code{file} and \code{local} arguments.}

\item{envir}{An environment to source into (by default the global
environment).}
}
\value{
The environment into which the code is sourced,
invisibly.  This is primarily useful when used as
\code{source_maker_script(script, envir=new.env())}, as the
environment created in the call is returned.
}
\description{
Convenience function for sourcing a maker script.  This just takes
care of writing the character vector to a temporary file and
running R's \code{\link{source}} over it.  It will also source
other arbitrary sets of R code that are character vectors rather
than files.
}

