% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{store}
\alias{store}
\title{Data store used by maker}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fd339ca9790> 
 - attr(*, "name")= chr "store_generator"
}}
\usage{
store
}
\description{
The data store used by maker.  This is an R6 class, and generally
does not need creating manually.  It's basically a pair of
key-value stores; one for R objects and one for files.
}
\keyword{datasets}

