% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse_command.R
\name{target_is_file}
\alias{target_is_file}
\title{Determine if target is a file}
\usage{
target_is_file(x)
}
\arguments{
\item{x}{Vector of target names}
}
\value{
A logical vector, the same length as \code{x}
}
\description{
Determine if a target is treated as a file or not.
}
\details{
A target is a file if it contains a slash anywhere in the name, or
if it ends in one of the known extensions.  The current set of
known file extensions is available as \code{\link{extensions}()},
but soon will become configurable.
}

