% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/use_remake.R
\name{use_remake}
\alias{use_remake}
\title{Add a remake file template}
\usage{
use_remake(dir = ".")
}
\arguments{
\item{dir}{directory in which to create. Defaults to
current directory.}
}
\value{
This function is primarily useful for its side
  effect, which is to add a template remake.yml file to
  the home directory.
}
\description{
This function adds a simple remake file template to a
target directory.
}

