% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{diagram}
\alias{diagram}
\title{Make a figure with the dependency graph}
\usage{
diagram(..., remake_file = "remake.yml")
}
\arguments{
\item{...}{Additional arguments that control formatting but aren't
documented and are subject to change.}

\item{remake_file}{Name of remake file (default is
\code{remake.yml}).}
}
\value{
An htmlwidgets object, which are embeddable in all sorts
of things and print to the screen and seem very fancy.  See the
\code{grViz} help for more information.
}
\description{
Plot the graph that remake generates.
}
\details{
This is really just a placeholder, but I want this here early as
an indication of where the package is headed.  Plus this is
something I have always wanted in \code{make}.  Current version is
not tunable on purpose.
}
