% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diagram.R
\name{diagram}
\alias{diagram}
\title{Make a figure with the dependency graph}
\usage{
diagram(remake_file = "remake.yml", ...)
}
\arguments{
\item{remake_file}{Name of remake file (default is}

\item{...}{Additional arguments thta control formatting but aren't
documented and are subject to change.}
}
\description{
Plot the graph that remake generates.
}
\details{
This is really just a placeholder, but I want this here early as
an indication of where the package is headed.  Plus this is
something I have always wanted in make.  Current version is not
tunable on purpose.
}

