% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fetch}
\alias{fetch}
\title{Fetch last computed}
\usage{
fetch(target_name, require_current = FALSE, verbose = TRUE,
  allow_missing_packages = FALSE, remake_file = "remake.yml")
}
\arguments{
\item{target_name}{The name of a single target to fetch the value
of}

\item{require_current}{Logical indicating if the targets must be
up-to-date to be fetched.  If this is \code{TRUE} and the targets
are not up-to-date, then an error will be thrown.}

\item{verbose}{Be verbose when loading remake file?  Default is
\code{TRUE}.}

\item{allow_missing_packages}{Allow missing packages when loading
remake file?}

\item{remake_file}{Name of the remakefile (by default
\code{remake.yml})}
}
\value{
An R object.
}
\description{
Fetch the last computed value from the remake database.
}
\details{
The last computed value would be returned invisibly by
\code{make}, but this function provides a way of accessing values
without ever triggering a rebuild.  As such, it's possible that the
target is not made, or is not current, so there are options for
controlling what to do in this case.

It is an error to use this function with file targets (but see
\code{\link{is_current}} for checking currentness) and
\code{fetch_archive} for extracting files from archives.
}

