% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{file_extensions}
\alias{file_extensions}
\title{Vector of file extensions}
\usage{
file_extensions()
}
\description{
Returns the vector of known file extensions.  If a target ends in
one of these, then it will be considered a file, rather than an
object.  In a future version, it might be possible to configure
additional extensions: please let me know if that would be useful.
}

