% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{install_missing_packages}
\alias{install_missing_packages}
\title{Install missing packages}
\usage{
install_missing_packages(remake_file = "remake.yml", instructions = FALSE,
  missing_only = TRUE, skip_target_packages = FALSE)
}
\arguments{
\item{remake_file}{Name of the remakefile to look for the list of
required packages}

\item{instructions}{Rather than install anything, prints
instructions on how to install missing things}

\item{missing_only}{If \code{FALSE}, install \emph{everything},
rather than just missing packages.  This might be useful to set to
\code{TRUE} in conjunction with \code{instructions=TRUE} to
generate a full list to install.}

\item{skip_target_packages}{Skip packages that are mentioned only
in targets?}
}
\description{
Install missing packages
}
\details{
No version comparison is done - see packrat for a more complete
package management solution, though in my initial experiments it
does not play that nicely with remake.
}
\author{
Rich FitzJohn
}

