% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api.R
\name{install_remake}
\alias{install_remake}
\title{Install running script.}
\usage{
install_remake(dest)
}
\arguments{
\item{dest}{Directory to install `remake` to.  Should be on your
path, though the current directory may be useful too.}
}
\description{
Install running script to a local directory.  This directory
should be on the \code{$PATH}.  Once this has been done, you can
run remake with `remake`.  See `remake --help` for more information.
}
\details{
The installed script is just a wrapper to the function
\code{\link{main}}; this means that upgrades to remake do not
require this to be rerun.  The installed script is extremely
simple.

Because of the design of \code{\link{Rscript}}, the helper script
loads the method package: in my experience many things just do not
work without that package loaded and Rscript does not load it by
default (especially in conjunction with \code{::}).
}

