% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remake.R
\name{is_current}
\alias{is_current}
\title{Check if a target is current or not}
\usage{
is_current(target_name, m = NULL, check = NULL)
}
\arguments{
\item{target_name}{Name of the target.  An error is thrown if the
target does not exist.}

\item{m}{A remake object.  If omitted, then one will be built using}

\item{check}{What to check.  It can be "exists", "depends", "code"
or "all.  By default, this comes from the target default.}
}
\description{
Check if a target is current or not.
}
\author{
Rich FitzJohn
}

