% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Command line interface to maker}
\usage{
main(args = commandArgs(TRUE))
}
\arguments{
\item{args}{Arguments to pass to}
}
\description{
The command line script, to be run outside of the package.  Don't
run this from within R (well, you can't actually).  This uses
\code{commandArgs} to pass along arguments to \code{\link{maker}}.
}

