% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make}
\alias{make}
\title{Make a single target}
\usage{
make(target, maker_file = "maker.yml", path = ".")
}
\arguments{
\item{target}{Name of a target to build}

\item{maker_file}{Name of the makerfile (by default
\code{maker.yml}).  This is passed to \code{maker$new()}.}

\item{path}{Path to build in (by default the current directory).
Probably safest to leave this as-is.}
}
\description{
This is a convenience function that creates a maker object and
builds a single target.  If that target is an object it will be
invisibly returned.
}

