% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api.R
\name{make}
\alias{make}
\title{Make a single target}
\usage{
make(target_name = NULL, ..., remake_file = "remake.yml", verbose = TRUE)
}
\arguments{
\item{target_name}{Character vector of names of targets to build,
or \code{NULL} to build the default target (if specified in the
remakefile).}

\item{...}{Additional future arguments, ignored for now.
Practically this means that all other arguments must be specified
by full name.}

\item{remake_file}{Name of the remakefile (by default
\code{remake.yml}).}

\item{verbose}{Controls whether remake is verbose or not.  By
default it is (\code{TRUE}), which prints out the name of each
target as it is built/checked.  This argument is passed to
\code{\link{remake_verbose}}; valid options are \code{TRUE},
\code{FALSE} and also the result of calling
\code{\link{remake_verbose}}.}
}
\description{
Make a single target
}

