% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make}
\alias{make}
\alias{make_script}
\title{Make a single target}
\usage{
make(target_names = NULL, maker_file = "maker.yml", path = ".")

make_script(target_names = NULL, maker_file = "maker.yml", path = ".")
}
\arguments{
\item{target_names}{Vector of names of targets to build, or
\code{NULL} to build the default target (if specified in the
makerfile).}

\item{maker_file}{Name of the makerfile (by default
\code{maker.yml}).  This is passed to \code{maker$new()}.}

\item{path}{Path to build in (by default the current directory).
Probably safest to leave this as-is.}
}
\description{
These are convenience functions that creates a maker object and
runs \code{make} or \code{script} on it.  For \code{make}, if the
final target is an object, that will be invisibly returned.  These
exist only to avoid creating a maker object before doing something
with it; this is all these functions do internally!
}

