% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api.R
\name{make_script}
\alias{make_script}
\title{Make a single target}
\usage{
make_script(target_name = NULL, verbose = FALSE, filename = NULL,
  remake_file = "remake.yml")
}
\arguments{
\item{target_name}{Character vector of names of targets to build,
or \code{NULL} to build the default target (if specified in the
remakefile).}

\item{verbose}{Be verbose when loading the remake file?}

\item{filename}{A filename to save the resulting script into.  If
\code{NULL} (the default) then an a character vector is returned
that can be inspected.  It can also be sourced without writing to
file using \code{\link{source_character}}.}

\item{remake_file}{Name of the remakefile (by default
\code{remake.yml}).  This is passed to \code{remake()}.}

\item{...}{Additional future arguments, ignored for now.
Practically this means that all other arguments must be specified
by full name.}
}
\description{
Make a single target
}

