% Generated by roxygen2 (4.0.2): do not edit by hand
\name{maker}
\alias{maker}
\title{Create a maker object}
\usage{
maker(maker_file = "maker.yml", verbose = TRUE, quiet_target = NULL)
}
\arguments{
\item{maker_file}{Name of the makerfile (by default
\code{maker.yml})}

\item{verbose}{Controls whether maker is verbose or not.  By
default it is (\code{TRUE}), which prints out what is going on.}

\item{quiet_target}{Controls whether maker quietens all targets by
default (default is \code{FALSE}).}
}
\description{
Creates a maker instance to interact with.
}

