% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{maker}
\alias{maker}
\title{Main maker object}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7f9dd4410190> 
 - attr(*, "name")= chr "maker_generator"
}}
\usage{
maker
}
\description{
The actual maker object to interact with.
}
\keyword{datasets}

