% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remake.R
\name{remake}
\alias{remake}
\title{Create a remake object}
\usage{
remake(remake_file = "remake.yml", verbose = TRUE, envir = NULL,
  allow_cache = TRUE)
}
\arguments{
\item{remake_file}{Name of the remakefile (by default
\code{remake.yml})}

\item{verbose}{Controls whether remake is verbose or not.  By
default it is (\code{TRUE}), which prints out the name of each
target as it is built/checked.  This argument is passed to
\code{\link{remake_verbose}}; valid options are \code{TRUE},
\code{FALSE} and also the result of calling \code{remake_verbose}.}

\item{envir}{An environment into which to create \emph{links} to
remake-controlled objects (targets and sources).  \code{.GlobalEnv}
is a reasonable choice.  This will change in a future version.}

\item{allow_cache}{Allow cached remake instances to be loaded?}
}
\description{
Creates a remake instance to interact with.
}
\examples{
\dontrun{
# create a quiet remake instance
m <- remake(verbose=FALSE)
# create a fairly quiet instance that does not print information
# for targets that do nothing (are up-to-date).
m <- remake(verbose=remake_verbose(noop=FALSE))
# Build the default target:
m$make()
}
}

