% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{remake}
\alias{remake}
\title{Command line interface to remake}
\description{
This function exists to support commandline use, as installed by
\code{\link{install_remake}}.  This little script accepts standard
unix-style command-line options to drive a subset of remake's
functionality.  Don't run this function from within R!  This uses
\code{commandArgs} to pass along arguments to various remake
functions.
}
\details{
See \code{remake --help} for instructions.

At the moment, supported featues are:
\itemize{

\item Running \code{remake::make}, as \code{remake target1
[target2]}.  If targets are ommitted the default target will be
used.

\item listing targets, as \code{--list-targets}, which calls
\code{\link{list_targets}}

\item generating a script with \code{-s} or \code{--script}
(printing to standard output) or \code{--script-file} prints to a
file.

\item return the version, as \code{-v} or \code{--version},
returning \code{packageVersion("remake")}
}

Additionally, the file used can be selected by using \code{-f} or
\code{--file} (following \code{make}'s convention), and remake can
be run with \code{verbose=FALSE} by passing in \code{-q} or
\code{--quiet}.
}

